// src/app/shared/interfaces/files_types.ts
export const file_types: {[key: number]: string} = {
  324: 'signature',
  325: 'identity_document_copy',
  326: 'rut',
  327: 'bank_certification',
  328: 'color_photo',
  329: 'arl_affiliation_certification',
  330: 'authorization_letter',
  331: 'commerce_chamber_certificate',
  332: 'identity_document_copy',
  333: 'confidentiality_agreement',
  334: 'data_update_form'
};

export enum FileType {
  AdditinalDocs = 345,
  ARLcertification = 329,
}


export enum OcFileTypes {
  SOCIAL_SECURITY = 347,
  ELECTRONIC_INVOICE = 348,
  TAX_AUDITOR_CERTIFICATE = 349,
  ARL_CERTIFICATE = 350,
  ANEXO = 351,
  INCOME_TAX_RETURN = 352,
  EXCEEDS_INCOME = 353,
  TAX_CONDITION = 354,
  MEDICAL_PREPAID = 355,
  HOUSING_CREDIT = 356,
  DEPENDENTS = 357,
  AFC_CONTRIBUTIONS = 358,
  VOLUNTARY_PENSION_CONTRIBUTIONS = 359,
  MINOR_CHILDREN = 360,
  CHILDREN_STUDY_CERTIFICATE = 361,
  CHILDREN_MEDICINE_CERTIFICATE = 362,
  PARTNER_MEDICINE_CERTIFICATE = 363,
  FAMILY_MEDICINE_CERTIFICATE = 364,
}

export enum InfoAdditionalTypes {
  INCOME_TAX_RETURN = 125,
  EXCEEDS_INCOME = 126,
  TAX_CONDITION = 127,
  MEDICAL_PREPAID = 128,
  HOUSING_CREDIT = 129,
  DEPENDENTS = 130,
  AFC_CONTRIBUTIONS = 131,
  VOLUNTARY_PENSION_CONTRIBUTIONS = 132,
  MINOR_CHILDREN = 133,
  CHILDREN_STUDY_CERTIFICATE = 134,
  CHILDREN_MEDICINE_CERTIFICATE = 135,
  PARTNER_MEDICINE_CERTIFICATE = 136,
  FAMILY_MEDICINE_CERTIFICATE = 137,
}